+++ src/main/java/com/multi/toonGather/social/controller/SocialController.java
@@ -3,8 +3,8 @@ package com.multi.toonGather.social.controller;
 import com.multi.toonGather.common.exception.AccessDeniedException;
 import com.multi.toonGather.common.model.dto.PageDTO;
 import com.multi.toonGather.security.CustomUserDetails;
-import com.multi.toonGather.social.model.dto.DiaryDTO;
-import com.multi.toonGather.social.model.dto.ReviewDTO;
+import com.multi.toonGather.social.model.dto.diary.DiaryDTO;
+import com.multi.toonGather.social.model.dto.review.ReviewDTO;
 import com.multi.toonGather.social.service.SocialService;
 import com.multi.toonGather.user.model.dto.UserDTO;
 import com.multi.toonGather.webtoon.model.dto.WebtoonDTO;
