+++ src/main/java/com/multi/toonGather/social/service/SocialServiceImpl.java
@@ -2,8 +2,8 @@ package com.multi.toonGather.social.service;
 
 import com.multi.toonGather.common.exception.NotFoundException;
 import com.multi.toonGather.common.model.dto.PageDTO;
-import com.multi.toonGather.social.model.dto.DiaryDTO;
-import com.multi.toonGather.social.model.dto.ReviewDTO;
+import com.multi.toonGather.social.model.dto.diary.DiaryDTO;
+import com.multi.toonGather.social.model.dto.review.ReviewDTO;
 import com.multi.toonGather.social.model.mapper.SocialMapper;
 import com.multi.toonGather.user.model.dto.UserDTO;
 import com.multi.toonGather.webtoon.model.dto.WebtoonDTO;
