+++ src/main/resources/static/css/likeButton.css
@@ -1,0 +1,24 @@
+.like-button {
+  font-size: 1.5rem;
+  line-height: 1.5rem;
+  cursor: pointer;
+  user-select: none;
+}
+
+.like-button input {
+  display: none;
+}
+
+.like-button label {
+  color: #ccc;
+  transition: color 0.2s ease-in-out;
+}
+
+.like-button label:hover,
+.like-button input:checked + label {
+  color: #ff4136;
+}
+
+.like-button label::before {
+  content: '\2665'; /* 하트 문자 */
+}
